Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@
 int main() {
     float a,b,c;
     scanf("%f %f %f",&a,&b,&c);
-    int sum = a+b+c;
+    float sum = a+b+c;
     float average = sum/3;
     printf("Average: %.2f",average);
     return 0;
